requirements.txt
playwright
pyyaml
Pillow